Switch the json sample's AST from scalajson to json4s
scalajson isn't published for scala 2.13. I had built and locally-published a
scalajson_2.13 build, which is how I was locally running the json sample with
scala3. However, the CI doesn't accept that as valid. Switching AST was the
easiest way to have the json sample work with scala3 without complicating the
build.