[base] Move typeclass implementations to be method-local
Apparently, removing a private inner class is a MiMa issue,
while removing a method-local class is not.

This moves towards the more future-flexible encoding.

Since the typeclass implementations are being moved anyway,
add apply methods to the typeclass companion objects and
use the apply in the companion object's other methods
to simplify the encoding